+++ server/publications/accountSettings.js
@@ -1,0 +1,3 @@
+Meteor.publish('accountSettings', function() {
+  return AccountSettings.find();
+});